SafetyBrain/requirements.txt
asgiref
certifi==2022.6.15
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django 
django-admin-rangefilter==0.8.7
django-filter==22.1
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
Markdown==3.4.1
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse
tablib==3.2.1
typing_extensions
tzdata 
urllib3==1.26.11
xlrd==2.0.1
xlwt==1.3.0
